requirements.txt
numpy
gunicorn
Flask
tabulate
networkx
mkdocs-material
mkdocs
anyio==4.4.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.1
filelock==3.15.4
Flask==3.0.3
fsspec==2024.6.0
gradio_client==1.0.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
packaging==24.1
pillow==10.3.0
PyYAML==6.0.1
requests==2.32.3
safetensors==0.4.3
sniffio==1.3.1
sympy==1.12.1
timm==1.0.7
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.1
websockets==11.0.3
Werkzeug==3.0.3

